Changed header file to take u as long double

+++ include/PsiTools.h
@@ -15,16 +15,16 @@ typedef long long integer;
 long double T(long double);
 long double psi_work(uint64_t);
 long double psi(uint64_t);
-long double S1(uint64_t, uint64_t);
-long double S2(uint64_t, uint64_t);
-long double S3(uint64_t, uint64_t);
-long long S3_B(uint64_t, uint64_t, uint64_t);
-long double S3_A(uint64_t, uint64_t, uint64_t);
-long double S4(uint64_t, uint64_t);
-long double S4a(uint64_t, uint64_t, long double);
-long double S4a_innerLoop(uint64_t, uint64_t, uint64_t, long double);
-long double S4b(uint64_t, uint64_t, long double);
-long double S4b_innerSum(uint64_t, uint64_t, uint64_t, long double);
+long double S1(uint64_t, long double);
+long double S2(uint64_t, long double);
+long double S3(uint64_t, long double);
+long long S3_B(uint64_t, long double, uint64_t);
+long double S3_A(uint64_t, long double, uint64_t);
+long double S4(uint64_t, long double);
+long double S4a(uint64_t, long double, long double);
+long double S4a_innerLoop(uint64_t, long double, uint64_t, long double);
+long double S4b(uint64_t, long double, long double);
+long double S4b_innerSum(uint64_t, long double, uint64_t, long double);
 // uint64_t pow(uint64_t, uint64_t);
 long long mobius(long long);
 long long mobius_work(long long);
@@ -32,8 +32,8 @@ void mobius_setup();
 void psi_setup();
 void setupEnvironment();
 
-long double slowS4(uint64_t x, uint64_t u);
-long double slowS4_inner(uint64_t x, uint64_t u, uint64_t l, long double psiOfU);
+long double slowS4(uint64_t x, long double u);
+long double slowS4_inner(uint64_t x, long double u, uint64_t l, long double psiOfU);
 #endif //PSITOOLS_H
 
 static B2Class B2;
